feat: split GSheets
