+++ setup.cfg
@@ -92,7 +92,7 @@ console =
 shillelagh.adapter =
     socrataapi = shillelagh.adapters.api.socrata:SocrataAPI
     csvfile = shillelagh.adapters.file.csvfile:CSVFile
-    gsheetsapi = shillelagh.adapters.api.gsheets:GSheetsAPI
+    gsheetsapi = shillelagh.adapters.api.gsheets.adapter:GSheetsAPI
     weatherapi = shillelagh.adapters.api.weatherapi:WeatherAPI
 sqlalchemy.dialects =
     shillelagh = shillelagh.backends.apsw.dialects.base:APSWDialect
