+++ src/shillelagh/adapters/api/gsheets/typing.py
@@ -1,0 +1,84 @@
+from typing import Any
+from typing import List
+
+from typing_extensions import Literal
+from typing_extensions import TypedDict
+
+
+class UrlArgs(TypedDict, total=False):
+    headers: int
+    gid: int
+    sheet: str
+
+
+class QueryResultsColumn(TypedDict, total=False):
+    id: str
+    label: str
+    type: str
+    pattern: str  # optional
+
+
+class QueryResultsCell(TypedDict, total=False):
+    v: Any
+    f: str  # optional
+
+
+class QueryResultsRow(TypedDict):
+    c: List[QueryResultsCell]
+
+
+class QueryResultsTable(TypedDict):
+    cols: List[QueryResultsColumn]
+    rows: List[QueryResultsRow]
+    parsedNumHeaders: int
+
+
+class QueryResultsError(TypedDict):
+    reason: str
+    message: str
+    detailed_message: str
+
+
+class QueryResults(TypedDict, total=False):
+    """
+    Query results from the Google API.
+
+    Successful query:
+
+    {
+        "version": "0.6",
+        "reqId": "0",
+        "status": "ok",
+        "sig": "1453301915",
+        "table": {
+            "cols": [
+                {"id": "A", "label": "country", "type": "string"},
+                {"id": "B", "label": "cnt", "type": "number", "pattern": "General"},
+            ],
+            "rows": [{"c": [{"v": "BR"}, {"v": 1.0, "f": "1"}]}],
+            "parsedNumHeaders": 0,
+        },
+    }
+
+    Failed:
+
+    {
+        "version": "0.6",
+        "reqId": "0",
+        "status": "error",
+        "errors": [
+            {
+                "reason": "invalid_query",
+                "message": "INVALID_QUERY",
+                "detailed_message": "Invalid query: NO_COLUMN: C",
+            }
+        ],
+    }
+    """
+
+    version: str
+    reqId: str
+    status: Literal["ok", "error"]
+    sig: str
+    table: QueryResultsTable
+    errors: List[QueryResultsError]
